+++ cvat-core/src/api.js
@@ -116,13 +116,32 @@ function build() {
                 return result;
             },
             /**
+                * Method allows to register on a server
+                * @method register
+                * @async
+                * @memberof module:API.cvat.server
+                * @param {string} username An username for the new account
+                * @param {string} firstName A first name for the new account
+                * @param {string} lastName A last name for the new account
+                * @param {string} email A email address for the new account
+                * @param {string} password1 A password for the new account
+                * @param {string} password2 The confirmation password for the new account
+                * @throws {module:API.cvat.exceptions.PluginError}
+                * @throws {module:API.cvat.exceptions.ServerError}
+            */
+            async register(username, firstName, lastName, email, password1, password2) {
+                const result = await PluginRegistry
+                    .apiWrapper(cvat.server.register, username, firstName,
+                        lastName, email, password1, password2);
+                return result;
+            },
+            /**
                 * Method allows to login on a server
                 * @method login
                 * @async
                 * @memberof module:API.cvat.server
                 * @param {string} username An username of an account
                 * @param {string} password A password of an account
-                * @throws {module:API.cvat.exceptions.ScriptingError}
                 * @throws {module:API.cvat.exceptions.PluginError}
                 * @throws {module:API.cvat.exceptions.ServerError}
             */