+++ cvat/apps/authentication/views.py
@@ -1,13 +1,10 @@
-
 # Copyright (C) 2018 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
 from django.shortcuts import render, redirect
 from django.conf import settings
-from django.http import JsonResponse
 from django.contrib.auth import login, authenticate
-from django.middleware.csrf import get_token
 
 from . import forms
 
@@ -24,7 +21,3 @@ def register_user(request):
     else:
         form = forms.NewUserForm()
     return render(request, 'register.html', {'form': form})
-
-
-def get_csrf(request):
-    return JsonResponse({'csrf': get_token(request)})