+++ cvat/apps/engine/urls.py
@@ -40,5 +40,6 @@ urlpatterns = [
     path('api/docs/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
 
     # entry point for API
+    path('api/v1/auth/', include('cvat.apps.authentication.api_urls')),
     path('api/v1/', include((router.urls, 'cvat'), namespace='v1'))
 ]